requirements.txt
# Core Framework
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
llama-index>=0.9.48
openai>=1.10.0

# Evaluation
deepeval>=0.20.0
ragas>=0.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Vector Database
qdrant-client>=1.7.0
sentence-transformers>=2.3.0

# Knowledge Graph
neo4j>=5.16.0

# Multimodal Processing
# Image
Pillow>=10.0.0
pytesseract>=0.3.10
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0

# Audio
openai-whisper>=20231117
pydub>=0.25.1
librosa>=0.10.0

# Video
opencv-python>=4.8.0
moviepy>=1.0.3
scenedetect[opencv]>=0.6.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.0.0
python-pptx>=0.6.21

# Text Processing
spacy>=3.7.0
nltk>=3.8.0
rank-bm25>=0.2.2

# Embeddings
fastembed>=0.1.0

# Observability
loguru>=0.7.0
prometheus-client>=0.19.0

# Web Framework
streamlit>=1.31.0
streamlit-agraph>=0.0.45
plotly>=5.18.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
tenacity>=8.2.0
httpx>=0.26.0
aiofiles>=23.2.0
tqdm>=4.66.0
python-multipart>=0.0.6

# Data Validation
jsonschema>=4.21.0
validators>=0.22.0

# Caching
redis>=5.0.0
diskcache>=5.6.0

# Security
cryptography>=42.0.0

# Development
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0
ipython>=8.20.0
jupyter>=1.0.0

# Additional dependencies
scikit-learn>=1.3.0

